Bug 6193 - Change in length of DatapathId of a switch
Dpid can go out of the Long range which results in NumberFormatException.
Hence the need to use BigInteger to hold the Dpid

Current patch accomodates Dpid to be of type BigInteger.

Author: Sai MarapaReddy <sai.marapareddy@gmail.com>

+++ applications/forwardingrules-manager/src/main/java/org/opendaylight/openflowplugin/applications/frm/FlowNodeConnectorInventoryTranslator.java
@@ -8,7 +8,9 @@
 
 package org.opendaylight.openflowplugin.applications.frm;
 
+import java.math.BigInteger;
+
 public interface FlowNodeConnectorInventoryTranslator {
 
-    public boolean isNodeConnectorUpdated(long dpId, String portName);
+    public boolean isNodeConnectorUpdated(BigInteger dpId, String portName);
 }
